utc and local wrap the new command and as such aren't entirely core
